+++ tests/corpora/test_update_usfm_parser_handler.py
@@ -1,11 +1,10 @@
 from typing import Iterable, List, Optional, Sequence, Tuple, Union
 
 from testutils.corpora_test_helpers import USFM_TEST_PROJECT_PATH, ignore_line_endings
-from testutils.default_paratext_project_settings import DefaultParatextProjectSettings
+from testutils.memory_paratext_project_file_handler import DefaultParatextProjectSettings, MemoryParatextProjectTextUpdater
 
 from machine.corpora import (
     FileParatextProjectTextUpdater,
-    MemoryParatextProjectTextUpdater,
     ScriptureRef,
     UpdateUsfmMarkerBehavior,
     UpdateUsfmRow,
